Move contrib in garage.contrib
This refactor moves all files inside contrib into the package
garage.contrib.
The corresponding imports of the moved files were modified as well.
