--- contrib/ros/envs/sawyer/__init__.py
@@ -1,3 +1,0 @@
-from contrib.ros.envs.sawyer.pick_and_place_env import PickAndPlaceEnv
-from contrib.ros.envs.sawyer.push_env import PushEnv
-from contrib.ros.envs.sawyer.reacher_env import ReacherEnv
