--- contrib/ros/robots/__init__.py
@@ -1,1 +1,0 @@
-from contrib.ros.robots.sawyer import Sawyer
