--- contrib/ros/worlds/__init__.py
@@ -1,2 +1,0 @@
-from contrib.ros.worlds.block_world import BlockWorld
-from contrib.ros.worlds.empty_world import EmptyWorld
