+++ garage/contrib/ros/envs/sawyer/__init__.py
@@ -1,0 +1,3 @@
+from garage.contrib.ros.envs.sawyer.pick_and_place_env import PickAndPlaceEnv
+from garage.contrib.ros.envs.sawyer.push_env import PushEnv
+from garage.contrib.ros.envs.sawyer.reacher_env import ReacherEnv
