+++ garage/contrib/ros/envs/sawyer/launchers/trpo_real_sawyer_push.py
@@ -1,10 +1,8 @@
-"""Example launcer file for experiment on real sawyer with trpo."""
-
 import numpy as np
 import rospy
 
-from contrib.ros.envs.sawyer import ReacherEnv
-from garage.baselines import LinearFeatureBaseline
+from garage.baselines.linear_feature_baseline import LinearFeatureBaseline
+from garage.contrib.ros.envs.sawyer.push_env import PushEnv
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
 from garage.theano.algos import TRPO
@@ -22,20 +20,20 @@ INITIAL_ROBOT_JOINT_POS = {
 
 
 def run_task(*_):
-    """Run task function."""
-    initial_goal = np.array([0.6, -0.1, 0.30])
+    initial_goal = np.array([0.6, -0.1, 0.80])
 
-    rospy.init_node('trpo_real_sawyer_reacher_exp', anonymous=True)
+    rospy.init_node('trpo_real_sawyer_push_exp', anonymous=True)
 
-    env = ReacherEnv(
+    push_env = PushEnv(
         initial_goal,
         initial_joint_pos=INITIAL_ROBOT_JOINT_POS,
-        simulated=False,
-        robot_control_mode='position')
+        simulated=False)
 
-    rospy.on_shutdown(env.shutdown)
+    rospy.on_shutdown(push_env.shutdown)
 
-    env.initialize()
+    push_env.initialize()
+
+    env = push_env
 
     policy = GaussianMLPPolicy(env_spec=spec(env), hidden_sizes=(32, 32))
