+++ garage/contrib/ros/envs/sawyer/launchers/trpo_real_sawyer_reacher.py
@@ -1,8 +1,10 @@
+"""Example launcer file for experiment on real sawyer with trpo."""
+
 import numpy as np
 import rospy
 
-from contrib.ros.envs.sawyer.push_env import PushEnv
-from garage.baselines.linear_feature_baseline import LinearFeatureBaseline
+from garage.baselines import LinearFeatureBaseline
+from garage.contrib.ros.envs.sawyer import ReacherEnv
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
 from garage.theano.algos import TRPO
@@ -20,20 +22,20 @@ INITIAL_ROBOT_JOINT_POS = {
 
 
 def run_task(*_):
-    initial_goal = np.array([0.6, -0.1, 0.80])
+    """Run task function."""
+    initial_goal = np.array([0.6, -0.1, 0.30])
 
-    rospy.init_node('trpo_real_sawyer_push_exp', anonymous=True)
+    rospy.init_node('trpo_real_sawyer_reacher_exp', anonymous=True)
 
-    push_env = PushEnv(
+    env = ReacherEnv(
         initial_goal,
         initial_joint_pos=INITIAL_ROBOT_JOINT_POS,
-        simulated=False)
+        simulated=False,
+        robot_control_mode='position')
 
-    rospy.on_shutdown(push_env.shutdown)
+    rospy.on_shutdown(env.shutdown)
 
-    push_env.initialize()
-
-    env = push_env
+    env.initialize()
 
     policy = GaussianMLPPolicy(env_spec=spec(env), hidden_sizes=(32, 32))
