+++ garage/contrib/ros/envs/sawyer/launchers/trpo_sim_sawyer_reacher.py
@@ -6,8 +6,8 @@ import moveit_commander
 import numpy as np
 import rospy
 
-from contrib.ros.envs.sawyer import ReacherEnv
 from garage.baselines import LinearFeatureBaseline
+from garage.contrib.ros.envs.sawyer import ReacherEnv
 from garage.envs.util import spec
 from garage.misc.instrument import run_experiment
 from garage.theano.algos import TRPO
