+++ garage/contrib/ros/envs/sawyer/pick_and_place_env.py
@@ -6,9 +6,9 @@ import gym
 import moveit_commander
 import numpy as np
 
-from contrib.ros.envs.sawyer.sawyer_env import SawyerEnv
-from contrib.ros.robots.sawyer import Sawyer
-from contrib.ros.worlds.block_world import BlockWorld
+from garage.contrib.ros.envs.sawyer.sawyer_env import SawyerEnv
+from garage.contrib.ros.robots.sawyer import Sawyer
+from garage.contrib.ros.worlds.block_world import BlockWorld
 from garage.core import Serializable
