+++ garage/contrib/ros/robots/__init__.py
@@ -1,0 +1,1 @@
+from garage.contrib.ros.robots.sawyer import Sawyer
