+++ garage/contrib/ros/robots/robot.py
