+++ garage/contrib/ros/worlds/__init__.py
@@ -1,0 +1,2 @@
+from garage.contrib.ros.worlds.block_world import BlockWorld
+from garage.contrib.ros.worlds.empty_world import EmptyWorld
