+++ garage/contrib/ros/worlds/models/block/model.urdf
