+++ garage/contrib/ros/worlds/models/cafe_table/meshes/cafe_table.dae
