+++ garage/contrib/ros/worlds/models/cafe_table/model.config
