+++ garage/contrib/ros/worlds/world.py
